Ensure "readonly" files can be written to.
If a "readonly" file has mode 0400, then the root user can write
to it regardless of the on-disk file permissions. However, if the
daemon is run unprivileged then the owner of these files is not
able to write to them -- instead, force the mode to 0600, which
is actually what we want.